Fix simulation of mulhw

+++ src/ppc/simulator-ppc.cc
@@ -1916,9 +1916,9 @@ void Simulator::DecodeExt2_9bit(Instruction* instr) {
       int rt = instr->RTValue();
       int ra = instr->RAValue();
       int rb = instr->RBValue();
-      int64_t ra_val = (get_register(ra) & 0xFFFFFFFF);
-      int64_t rb_val = (get_register(rb) & 0xFFFFFFFF);
-      int64_t alu_out = ra_val * rb_val;
+      int32_t ra_val = (get_register(ra) & 0xFFFFFFFF);
+      int32_t rb_val = (get_register(rb) & 0xFFFFFFFF);
+      int64_t alu_out = (int64_t)ra_val * (int64_t)rb_val;
       alu_out >>= 32;
       set_register(rt, alu_out);
       if (instr->Bit(0)) {  // RC bit set
